Move BSF from InternalMath to InternalBit
